requirements.txt
torch~=2.2.2
torchvision~=0.17.2
matplotlib~=3.8.4
pillow~=10.3.0